move the title I county file to title I folder
