+++ src/components/title1/Title1CountyMap.tsx
@@ -25,7 +25,7 @@ import {
     getTitle1CountyScopedRecord,
     Title1CountyColumnConfig,
     Title1CountySelector
-} from "../../utils/title1County";
+} from "./title1County";
 
 interface Title1CountyMapProps {
     title: string;
